backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
eventlet==0.33.3
bcrypt==4.0.1
psycopg2-binary==2.9.9
Werkzeug==2.3.7
pytest==7.4.3
flask-swagger-ui==4.11.1
apispec==6.3.0
marshmallow==3.20.1